Site Overhaul: Simplification
Massive update today and a lot to share:
- Site has been overhauled (and by that I mean completely rewritten and
only assets from the previous version were taken over again.)
- Switch to Open Sans
- All sizes should be mobile compatible (em measurement used)
- Gazebo (large version) profile image used on index.
- Top-down theme instead of center-focus
- New nav bar overhaul
- /me page rewritten and closer to resume
- /connected renamed as /contact
- New dark social icons
- New Dark Mode support (based on user’s location between the hours of
10 PM and 5 AM)
- New main.css page to remove as much HTML styling as possible. This is
the default stylesheet when JavaScript is disabled in the user’s
browser.
- New dark.css page added, which Dark Mode will take advantage of
during its operating hours.
- Link underline removed and custom 1-pixel padding line added (due to
issue with the way default underline appeared)
- /lastnamevideo redirect page removed
That’s about it. Also, thanks to Tyler Eich for a bug fix in the
JavaScript. Today I learned that each function needs to be unique
within the directory.
Eat cake and prosper.