Mobile bug fix
Bug fixes for mobile again. Once again, line height was the issue and I
made up for its buffer room by adding a margin. Also, the image on the
index page caused horizontal scrolling, so it was made mobile friendly
by making the actual size of the image a width of 500pixels and setting
the max-width 100% with auto height. Lots of learning today.