Bugfixes
CSS needed adjusted for nav bar because it looks great on desktop but
not a bit close to the top on mobile devices. Hopefully this fixes it.
Picture on index page looks great on Mac and iPhone but giant on iPad.
Maybe max-width will fix this. Testing with this commit.